requirements.txt
numpy
bleak
face_recognition
requests
Flask==2.3.2
pymongo==4.5.0
python-dotenv==1.0.0
bson==0.5.10
portalocker==2.8.2
